rename: hooks하위 custom, query로 분리 (#115)
