+++ src/pages/AddForFindGuide.tsx
@@ -3,7 +3,7 @@ import Introduction from '@/components/addTravel/Introduction';
 import ScheduleTeam from '@/components/addTravel/ScheduleTeam';
 import Thumbnail from '@/components/addTravel/Thumbnail';
 import GrayBack from '@/components/GrayBack';
-import useHandleImageUpload from '@/hooks/useHandleImageUpload';
+import useHandleImageUpload from '@/hooks/custom/useHandleImageUpload';
 import { addTravelWrapper, noneStyleInput, pageLayoutWrapper } from '@/pages/AddTravel';
 import useImageStore from '@/stores/useImageStore';
 import useSectionsStore from '@/stores/useSectionsStore';
